analysis: add scifact first-stage diagnostic
Refs #188.
